requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
geopy

streamlit

requests

python-dotenv
